requirements.txt
urllib3
requests
discord.py
numpy
pyaudio
Pillow
gTTS
pygame
python-dotenv
opencv-python
pycaw
comtypes
